Add team to get_monitoring_stanza and get_paasta_config.
This is not yet hooked up to main() and command-line args; I'm going to take a detour to look at argparse at kmosher's suggestion
as he says it solves removes manual option validation drudgery.

This is probably going to end up as a separate --team and --notification-email but I'll start with TSTTCPW until we figure out
exactly what we want to do with this.