get_container_name()
We need a way to find a container that we launch in the background. We can't use image id
because there could be multiple containers running from that id. We could parse the output
of docker run (it emits the id of the container it launches) but I'm wary of cleaning up
and parsing that output.

Besides, passing a human-readable name is handy anyway (easy to find your test_run-launched
containers in docker ps). And once we know the name, we can figure out the container id.

Also I want this commit message to contain the phrase 'a priori' so now it does.