More helpful output for interactive use.
I think kwa will argue this script is for machines so the interface isn't as important.
I got here because I wanted to be able to test this fellow so I needed some custom
handling so I decided to go ahead and add some info for humans just in case.