Fixed the weird /usr/bin/docker path manipulation that I copied from the old code and this seems to work.
Fixing that was enabled by paying attention to returncode from _run().

It also made my get_container_id() code work, so I'm going to continue to not parse actual
output from _run(). I think launching the container and then actually finding it
is a nice proof that the container is actually started. I'll update the error
message to better reflect the situation.