Exit with status code 2 on failure conditions.
This is the beginning of work to make paasta_metastatus more
useful(PAASTA-772).

This forces paasta_metastatus to exit with status code 2 when it
encounters certain failure conditions, namely: when the mesos master is
unavailable, when the number of masters, when the number of masters in
the cluster is lower than the configured quorum size, or the number of CPUs
available falls below 10% of the total.