remove use of exceptions as control flow
rather than raising exceptions to indicate unhealthy state, return
a tuple with each health check, giving a formatted message and a boolean
indicating health. The only exceptional circumstance is now when the
mesos master is completely unavailable, in which case we just bomb out
early.

this will cause exit status 2 on the following checks:
  - mesos master not reachable
  - no mesos quorum
  - less than 10% cpus available
  - less than 10% memory available
  - no marathon apps

If the mesos master is unavailable, then it will exit instantly.
Otherwise, the program will still print output for the other checks.
Other information printed includes:
  - number of mesos tasks
  - number of mesos slaves
  - number of marathon deployments
  - number of marathon tasks