Multiple jobs is not an error
I'm not really sure what the plan is for naming and how many jobs we need to keep to
retain history. For now, let's get status for all the jobs we know about. I predict
it will at least be useful early on for debugging. We can polish it out later if it
outlives its usefulness.