Delete get_matching_jobs. Replace with calls to newly improved lookup_chronos_jobs.
This is a long-delayed repsonse to kwa and robj comments in
review:117855. In my analysis, the underlying theme in their complaints
is that lookup_chronos_job is designed wrong. Correcting this malady is
what has motivated my last several refactoring branches: making
[de]compose_job_id more sensible, reducing the number of different ways
to accomplish various common job_id transformations, etc.