Extract filter_chronos_jobs from lookup_chronos_jobs.
This was after discussion with robj in
https://github.com/Yelp/paasta/pull/24 and then in person.  Basically
this is a compromise between SRP and DI (which favor calculating the
list of jobs separately from filtering it) and YAGNI (currently every
caller of lookup_chronos_jobs would need to do its own client.list() and
pass that in). Hence, filter_chronos_jobs is fully SRP and DI while
lookup_chronos_jobs is less pure but more convenient.