add a 'jobs' dict to itest context
currently, the itests only operate on a single job at a time,
hardcoding the reference in the steps. this adds a 'jobs' dict
to the context, to that the features can store references to
and work on multiple jobs at a time.

this also removes the steps that operated on ChronosJob objects,
given that they shouldn't really be used.