Fix command quoting issues
The command the container is supposed to run is being passed to
pipes.quote() once in get_docker_run_cmd() and once again in
run_docker_container().

If the string contains anything that pipes.quote thinks should be
quoted then we end up with some very weird results. For example
"echo 'Hello' && sleep 60" turns to "echo foo '"'"'&&'"'"' sleep 60".

This gets weirder when the command passed to us is already in the form
'''/bin/sh -c '/bin/echo "docker docker docker"' ''' because then we
prepend /bin/sh in front of it and quoting gets *very* ugly.