Handle target capacity as integer
AWS SFRs allow instance weights as floats but target capacity has to be
an integer. This ensures that we leave instance weights as floats when
we scale down but we then round to the nearest integer when setting the
target capacity.

It is recommended to weight your instances as integer values or accept
the rounding error when setting target capacity (which might be small in
a sufficiently large cluster).