Change to using docker-compose.yml v2
This changes the example_cluster to use an override on the itest docker
compose. To support the depends_on keyword we need to move to v2 of the
docker-compose file. This changes the way networking works and requires
an upgrade to docker-compose. The new networking no longer requires
explicit linking of containers.