Fix itests orphaned containers bug.
After this change in Mesos: https://reviews.apache.org/r/45454/ the
mesos agent will atempt to kill containers that it sees were started by
mesos but that it is not aware of starting. This is bad if you are
running one docker daemon with multiple agents (like our itests).
Luckily they gave us a flag to revert to the original behaviour.