check validity of clusters in start/stop/restart
rather than failing silently, exit early when the user specifies
an invalid cluster. An invalid cluster is one which the service
specified in the command is *not* deployed to, or, if the user
specifies an instance in the command, an invalid cluster is any
cluster which that service/instance is not deployed to.