Add caching on frameworks endpoint in mesos-cli
After 3d07cbeacc4cfe0e9556500a1e51cbe4b49e5f4c we are getting the
frameworks from the proper endpoint. However, we aren't caching the
response as we were with the state endpoint. This means that anywhere
which calls the master.framework (which includes things like
task.framework) had to make a new HTTP call to /master/frameworks each
time.

This manifested in slow runs of the autoscaler where we happen to call
task.framework once per task in the cluster. See
get_mesos_task_count_by_slave in mesos_tools.

This adds the same caching as is used on the state.json endpoint.