added request caching for service autoscaler
When using the mesos_cpu decision policy slave/state.json is queried for
each task in the service when calling slave.task_stats. These queries
are only cached on a per-task basis, instead of between tasks. This
results in a stave's state.json being queried for every task running on
it. Adding request caching in PaaSTA lets us cache the results between
tasks.