Fix waiting on mark for deployment
We introduced a bug in https://github.com/Yelp/paasta/pull/837 where
we would exit from wait for deployment after the first check of all the
clusters and instances.

This adds back the while loop which is neccessary even with the `with
progressbar` and also improves the test case which would have caught
this regression.