Fix ASG delta calculation for 0 capacity
ASGs can have 0 capacity, unlike SFRs. This change makes it so that we
try to launch an instance if we currently have 0 desired, unless this
violates max/min capacity. We need at least one instance to make a sane
calculation of utilization.

Also remove some log messages that can be confusing.