Get list of marathon apps once
We currently call marathon for a list of apps for every service and then
just filter the apps we're interested in.

This splits the getting of the apps from the API and the filtering into
separate functions so we can get the list of apps early and re-use that
list when we filter for each service/instance iteration. The aim here is
to speedup the overall bounce.