Fix mesos tools task counting
This changes the function that counts the number of tasks on a slave to
include any orphaned tasks. The orphaned tasks are still using
resources. Before this slaves with orphaned tasks would often report to
be running 0 and so the autoscaler would happily terminate them.

Orphaned tasks can occurr when a framework deregisters and can't
successfully tell mesos that it has deregistered.