Fix resource counting in metastatus
This changes the way we count cluster resources to include the resources
used by orphaned tasks.

Orphaned tasks can still be running and so are certainly still using
resources. This particularly affects the numbers returned to the
autoscaler so that it will take count of orphaned tasks.

Orphaned tasks can occur for example if marathon fails a leader election
or is stopped without being able to report it's status to mesos.