Bounce oracle change
This adds a new option to list_marathon_instances which lists all the
instances in one go but puts the instances expected to bounce first.
This way we can have one cron job (and so no race conditions) whilst
still speeding up the bounce by starting with the instances we expect to
bounce.