Move bounce lock to higher level
Currently we lock the do_bounce function on the service instance we are
bouncing. This moves the lock higher up to lock the `setup_service`
funcion. This should mean there can be no race conditions between two
instantiations of setup_marathon_job.

Note that this also makes failing to get the lock for a service a
non-fatal error. i.e. we just log and carry on. This is neccessary
because we sometimes expect two bounces to run at around the same time
so skipping one because it is in progress else where isn't really a
problem.

The motivation for this fix is that currently two invocations of
setup_marathon_job can gather the same info about a service and both
decide they need to create a new marathon app. Although only one gets a
lock to do so the second may eventually get the lock and gets an error
from marathon when it tries to create an app that doesn't exist.