Relax service autoscaler task counting
This makes two changes:
* Use the long marathon app_id. This means we should now only be
autoscaling the latest version of a paasta service. It also means we
can ignore any interaction with the bounce becase we are ignoring the
old service.
* Assume missing health checks are "normal" for tasks that haven't just
started. This is due to a bug in marathon where tasks can stop being
healthchecked but remain running and healthy.