Include orphan tasks in autoscaling decisions
This makes a simalar change to the service autoscaler that we have made
to the cluster autoscaler already in #968

The logic is that we never want to make autoscaling decisions where we
don't include all the tasks in the calculation. Even if a leader
election is in progress this should now count all the tasks.