Try multiple haproxy endpoints during bounce
Currently we connect to one smartstack endpoint to calculate which tasks
are happy for a particular service in a particular namespace. If the
host we've picked happens to have been terminated or is unreachable then
we error. This is slowing down the bounce process.

This commit makes it so that we keep trying any other relavent smartstack
host and only thow an error if we can't connect to any of them.

Internal ticket PAASTA-8357