Handle create app lock failure gracefully
We get these exceptions pretty frequently in our large clusters when we
are creating multiple apps at the same time. Currently the exception is
raised all the way up which causes setup_marathon_job to exit 1 and for
our deploy loop to start over. This catches the exception and continues
on. I think it should be safe because we will simply try again next time
to create the new marathon app.