Catch 409 conflict when bouncing marathon apps
In 0c950d0e43434222b598d8b8fc7da8c1383a666e we moved the lock in
`setup_marathon_job` to prevent race conditions between multiple bounce
loops which run at the same time. One motivation was seeing HTTP 409
conflicts from marathon when the bounce attempts to create the same app
twice. It turns out that this is still possible since we
`get_all_marathon_apps` outside the lock which can lead to two bounces
deciding to create the same app, resulting in HTTP 409 from marathon.

We could move this call into the lock but this would incurr a
performance hit as it would be called once per service instance. Instead
I am catching the error which I am pretty sure is safe. Essentially a
second bounce may decide it needs to create an app that already exists
but this is now a silent failure and the bounce will stop there.