Send git refs as byte strings
After we changed to decoding git refs to unicode strings we also need to
make sure to encode back to bytes before we send to dulwich. Otherwise
we see: refname is not a bytestring: u'refs/tags...'

I've also converted the shas back to bytes although I'm not sure we
really needed to...

I'm going to try and add some itests that will catch bugs like this in
the future. But I'll do that seperately.