Add optional autoscaling info to metastatus
This adds optional info on the cluster autoscaler to metastatus. It
works by using the autoscaling library to retrieve details on the
resources it is responsible for and contacting AWS to find the current
capacities of the resources. The motivation is so an operator can
easily see the settings relavent to the autoscaler and the current
capacity. Fixes #984

It is behind a command line flag so should be safe to ship as is. At
yelp this will need a wrapper script on the master that ensures the
autoscaling configs are synced from S3 and that there are AWS
credentials in the environment.