add contrib script for calculating utilization
I wanted a script that gave me an easier entrypoint for finding
utilization data of arbitrary sets of hosts in the cluster than
metastatus can provide. This is pretty simple: reads a list of hosts
from stdin and prints a json dump of total, used, perc for the list
of inputs. I can put this behind the api next, but this'll do for now