kill tasks that don't leave TASK_STAGING
create a timer that tracks a task when it is launched. if it does not
get to a running state within timeout, then kill the task. we've often
found problems when docker is broken on a mesos agent wherein a task
will forever be stuck staging and hangs forever more. adding a timer
that kills the task if it does not get launched within a given timeout
should prevent users forever waiting for their task to finish