Only update deployments.json if changed
This is a precursor to the work on paasta-deployd. It means that
deployments.json files will only be written to if they are being
updated. This makes it possible to watch for inotify updates to that
file and perform a bounce based on inotify events.

This also changes the cleanup behaviour for old deployments.json as we
can no longer rely on the timestamp to indicate a deleted service. Now
we check via paasta list.