"Fix" the itests
So this is pretty embarrasing but I added this test and it doesn't catch
the SystemExit that is called by setup_marathon_job. As a result our
itests are exiting whenever this test runs but seeming to be "green"
because the exit is with a 0 error code :-(

Merging this will "break" the build on master but in reality it's been
broken for quite a while!