Fix threaded timeout decorator
Currently the TimeoutError is thrown but the main thread will still
continue to wait for the function thread to finish before it exits
itself. Setting the daemon flag means that the child thread will be
killed if the main thread dies, which it does on receiving the timeout
exception.

Internally: PAASTA-10556