Add metric emission to deployd
This adds some basic metric emission to deployd. Various timers are
emitted by the deployd workers. In addition, we check the length of the
queues every 20s and emit a gauge.

I'm using yelp_meteorite because it is super easy for this purpose.
Unfortunately it's not public so I've hacked the import to fall back to
a mock and that will make the metrics code "do nothing" if you are
outside yelp. I discussed making this a pluggable interface with robj
but I think it's overkill to do so but please shout up if you think we
should.