more work on creating public interface
more work to address structure of program, creating the concept of a
Runner: an interface for running tasks, with the expectation that
implementations of runners will provide different concurrency
primitives.

Further reduce the interface provided by the Executor such that is is
just a wrapper around creating the execution framework and delegating
tasks to it. The Executor provides the public facing 'run' function -
this in turn delegates to the Runner chosen.