Make deployd check marathon ID before bouncing
As with the public config it doesn't make much sense to add things to
the bounce queue if their sha has not changed. This adds the same check
to changes to marathon-* files and deployments.json.

In addition I have disabled the yaml cache on the service configuration
lib since it was interfering with checking the sha. I'm pretty sure
we're not reading the configs frequently enough to need a cache.