Deployd actions for dying threads
This adds some sensible defences for workers and watchers dying due to
unexpected exceptions. In the case of a watcher dying we commit suicide
and simply let another leader take over. If a worker dies then we just
restart it, unless all workers die in which case we also commit suicide.