Make caching of marathon service config optional
This was confusing deployd and making it think it had reached a "steady
state" when infact a new app was desired. I think this is safe to
disable in setup marathon job because it's only called once per run so
the caching isn't really giving us any benefit there anyway.