Fix race watching new folder
From working on the itests I think there is a subtle race condition when
watching a new folder. If you watch the folder and then the files are
created after the watch is in place you end up bouncing a service. If
the files get placed before the watch is completely set up then they
aren't "new" to the watcher and so nothing is triggered. This commit
should trigger bounces/checking on new folders that have config files in
them.