Skip check for existing image in example cluster
We can't do this check in the example cluster because the registry URL
is "wrong". The registry has to be listed as localhost because it is
local to where the daemon is running. However localhost from the cli
tools does not resolve to the docker host and so we get a connection
error. Skipping the check with --force works fine.