Catch any exception running setup_marathon_job
I don't really like catching all exceptions but in this case it seems
reasonable just because it's hard to know everything that could go wrong
in setup_marathon_job, especially as it has to talk to marathon's API
and could end up getting a weird response.

The motivation is to prevent a service not being bounced again when
there happens to be some transient exception thrown in the worker.