Fix deployd not scaling when instance count change
When deployd notices a change to marathon-*.yaml it was checking if the
app ID had changed before adding a service to be bounced. Of course if
just the instance count has change the app ID is unchanged and nothing
happens.

This makes the code that checks the service needs bouncing more like the
oracle code in list_marathon_instances. There is a difference though, in
that it only checks the apps that it has been passed, i.e. those that we
know have had a change to their config file.