Ask marathon to embed tasks not failures
For some calls to /v2/apps we need to embed the tasks in the results. I
don't think we need the task failures and embedding them is deprecated
because it implies embedding everything else. See:
https://mesosphere.github.io/marathon/docs/rest-api.html#get-v2apps

Unfortunately embedding tasks still implies embedding deployments which
I think we also don't need. However, this is supposed to change at some
point in the future.

If the itests pass then I think this is safe and it should reduce the
load on marathon slightly.