Catch InvalidJobNameError in deployd
I'm not 100% on the root cause of this exception but the deploy daemons
file watcher certainly shouldn't die as a result. The safest thing is to
catch it and assume the service needs a bounce. Then it will be up to
the worker to have a go at deploying the service and back off it sees
the same error.