Dedupe the bounce queue
If a service is in the backlog to be bounced and there is a change to
its marathon-.yaml or deployments.json the the deploy daemon will notice
the change and if the new app that is required is not yet created then
it will add a second instance to the bounce queue. This way the queue
keeps growing.

This makes the BounceQueue drop any duplicate service instances that are
already in the queue.