Fix scaling SFRs to less than 1
When scaling SFRs we have to always kill one instance when down scaling.
This prevents us getting stuck when the non integer value of capacity
would take us below our integer target.

However, we can't set a capacity of 0 because it is not allowed by AWS
so we still need to prevent the autoscaler attempting to set a target of
0 even when it means it kills 0 instances. In addition, we must kill the
instance if the SFR is cancelled, it is safe to do so because we don't
need to set a target capacity on a cancelled SFR.