De-prioritise startup/public config bounces
This makes deployd workers pick bounces triggered on startup of deployd
or on changes to public config after all other bounces are done.

The motivation is to stop "big bounces" tying up workers that could be
dealing with real service pushes or autoscaling events. A "big bounce"
is one in which a config change or a code change in paasta_tools means
that we need to recreate every marathon app.