Make cleanup maintenance pass if unreserving fails
Currently we giveup undraining if we can't unreserve resources. I think
maybe this can happen for old hosts that have gone away or in other
circumstances. Either way we try again to unreserve later so it is not
essential that we fail early in this case. The downside of failing here
is the cleanup maint script may never finish.