Modify setup_marathon_job to work with sharded marathons.
This breaks deployd, which will be fixed in a future commit.
This also introduces stubs for the marathon-python library, so that mypy can better type-check calls to it. I generated the stubs with stubgen, and partially filled in the types of MarathonApp's parameters and member variables.
This also refactors bounce methods to take a set of opaque task objects, rather than dictionaries of app name -> task objects, since all of the bounce functions immediately consolidate all the tasks into a single list/set anyway.