Set new terminate timeout based on the number of tasks on killed agent
When scaling down a cluster, we can set the timout for the next slave
based on the running tasks on the most recently killed slave.  If, on
one terminate, we hit the timeout, this will try to set a lower timeout
for the next terminate based on the running tasks on the just killed
slave.