Added filter  for bogus mesos CPU values
The filter make sure the value reported by Mesos doesn't exceed what the
theorical maximum allowed CPU usage for this service.
THe feature is toggled by a system config. It's disabled by default so it
should be safe to roll out.
A Warning message will be written in the logs everytime such a value is
encountered. This debug verbosity, the entire CPU stats will be printed.