Bump drain timeout for hacheck
This is just a quick patch, we should refactor to make this configurable
really. The problem this helps with is when the deployd queues get large
it is possilbe that a draining task goes past its draining timeout
before it is re-processed. At that point we are in an infinite loop of
draining and undraining and things never get better. So this isn't a fix
for performance issues but does move the cliff edge a bit further back.