Enable yaml-cache in deployd
This re-enables the yaml cache in deployd whilst strategically clearing
the cache when deployd needs to be sure it is reading the latest data
from disk.

I'm pretty sure I caught all the places where I need to clear this
cache. Specifically, I've done anywhere in a watcher where the watcher
is using those yaml files to make a decision. I've also cleared it in
the worker before the worker runs SMJ for that service.