Fix randomly closing ZK connections in paasta-api
We've been getting exceptions from paasta-api about closed connections.
The problem occurs when there are multiple requests made to the
instances endpoint at the same time. For autoscaled services we use
ZookeeperPool to pool our connections to zookeeper so that we can get
the correct number of instances. However when the API handles multiple
requests it will call `.stop()` and `.close()` on the zookeeper client
when it is possilbly still being used by another request. This leads to
exceptions like `kazoo.exceptions.ConnectionClosedError`.

I've "fixed" this by adding the `ZookeeperPool` to the api main func so
that we don't tear down the client until the whole api server quits.