Add --skip-secrets option to local-run
For instances where we don't have a vault token but need to run
local-run non interactively it is useful to have an option that skips
decrypting secrets so that the non-interactive script isn't prompted for
a password.

This adds that option. The downside is that you just get the secret ref
left in the env var and not the plaintext.