[RFC] paasta api authentication
This is quite rough around the edges (or in other words I'm not super
happy about the implementation in its current form). Lacking tests etc.
I'm just opening this to get some opinions and see if anyone thinks it
is worth tidying up/working on for real.

The rough idea is to include a vault token in requests to the paasta API
as a header. The server then validates that the token is legit with
Vault and checks what policies are attached to the token. These vault
policies are then mappped to permissions on each api view.

I've also wrapped the client to generate + send the correct token in
response to 403s.