Prevent autoscalerwatcher adding services on start
When deployd starts it must add watches to all the zookeeper nodes for
autoscaled services. This also enqueues the services. This is normally
fine except in a "big bounce" scenario in which we don't want all
autoscaled services enqueued at startup.

This change passes an enqueue paramater to the watcher functions so that
new nodes that are created are enqueued for a bounce but nodes that are
enumerated on startup are not.