Try getting children from zk directly
I'm pretty sure this fixes a race condition in AutoscalerWatcher.

The race is between us setting a watch to pickup new child nodes and
using the list of current nodes in order to set watches on those too. If
the list of current nodes is retreived before we set the watch then in
theory a new child could appear before the watch is established and not
ever get a watch on it.

I can't see for sure in Kazoo the order but running this branch against
travis the itest for autoscaling passes every time whereas it fails ~50%
of the time on the previous commit. I think this will fix the cases
where we see deployd failing to scale newly enabled autoscaled services.