Fix another race in the AutoscalerWatcher
This makes it so the autoscaler watcher explicitly creates a node watch
on keys called "instances". Before this would only be the case if the
parent folder and instances key were created at nearly the same time.
i.e. the instances child existed by the time the watcher was being
created for the containing folder. Note that the added itest was failing
before I made this change.