Make chronos bounce secret aware
The bounce code needs to know about the signatures of secrets in the
secret json files. Otherwise the "config hash" doesn't change when the
plaintext of the secret changes because the environment we pass to
Chronos is just a reference to the secret.

This already exists for the marathon bounce, I've refactored that a bit
to maximise the code shared between the two.