Split out MAX_CLUSTER_DELTA for cluster autoscaler
I'm splitting this into a maximum increase and maximum decrease so that
I can reduce the maximum decrease without affecting the autoscaler's
ability to scale up.

The motivation is that because we now have multiple SFRs to speed up
scaling down we end up decreasing by 20% on each one and because we
round down to the nearest whole number we are effectively being more
aggressive than before both in speed and amount that we scale down.
So I've reduced the default max decrease to 10% and made it
configurable via /etc/paasta.