APOLLO-652: Make get_containers_and_ips.py compatible with tron/batch services.
Since tron/batch services do not supply MESOS_TASK_ID in their docker label
(for obvious reasons), we are updating get_containers_and_ips to use
only the paasta_service and paasta_instance in the mapping generated.

Since we are using only the service and instance fields of the
MESOS_TASK_ID (of otherwise marathon services), this should suffice.