Add config flags for get_x_running_here_for_nerve
I'm adding these so that we can disable registration of services on a
per node basis. For example, on k8s nodes we need to disable checking
for the mesos agent for services since the agent doesn't exist. And
vice-versa on k8s agents. We could try to make the code tolerate not
being able to reach the various agents but I think that would be bad
since it is probably by design to have configure_nerve crash if it can't
reach an agent rather than returning an empty list and de-registering
all the services.