Add cpu + mem constraints to k8s
Easier that I thought, we use the same values we use for marathon,
limits are "hard" requirements are "soft" but used by the scheduler.

I've tested that the code works and creates a deployment with the limits
on the container but haven't tried to actually push a container to see
what happens when you hit the limit.