Add support for StatefulSets
This gives the k8s logic the ability to manage StatefulSets. We manage
them the same way as Deployments. The distinguishing factor is that they
support dynamic EBS volumes that follow the Pods. The new yelpsoa syntax
looks like:
```
persistent_volumes:
  - container_path: /mnt/whatever
    size: 10
    mode: RW
```

Note this is deliberately a bit MVP. The StorageClass has to be created
by hand and we don't enforce constraints on AZ yet so you must be
careful to create everything in 1 AZ.