Ignore `kind` and `version` when calculating config hash
We're using `kind` and `api_version` to calculate the config hash.  Any
of these attributes could be changed during migration to a new version
of `k8s` client library, so we'll restart all instances as a result of
such change.

At the same time, these attributes are transport-only and are used only to
inform the `k8s` server endpoint about a class and a version of the class used
to generate JSON.  A change in these attributes without a change in other
attributes is no-op.

Thus, we better ignore changes in these attributes while calculating the config
hash.