Add deployd leader election check
Small script to check the number of contenders for a deployd leader and
alert if less than a certain number. We've seen a couple of cases where
a bug can cause the python processes to hang even though they've given
up leadership. In this case we might not have anyone left to take over
and this alert should let us know.