Also refactor SmartstackReplicationChecker
This makes SmartstackReplicationChecker the base class and pools the
logic that is just relevant to smartstack and paasta with a new sub
class MesosSmartstackReplicationChecker which contains the Mesos
specific agent filtering.

This is so I can create KubernetesSmartstackReplicationChecker which
will be basically the same but will "choose" agents to check based on
k8s labels rather than Mesos attributes.