Add `cleanup_self` to `@async_ttl_cache` and use it for all methods
`@async_ttl_cache` decorator caches a function call result using a key
composed from function's args values.  If it's used for a method, it
will use `self` as a part of the key, creating a strong reference to the
class instances in the internal cache as a result.  After some time,
references to the class instances are removed, so the class instance
reference in the cache becomes the only existing reference to that class
instance, and since the method is never called with that class instance
as a `self` arg, the corresponding record in the internal cache is never
expired, creating the memory leak as a result.

Introducing `cleanup_self` arg to `@async_ttl_cache` decorator, so while
it's `True` the decorator creates a weak reference to `self`, and
autocleans corresponding cache record on the object finalisation.  Using
it for all methods, so there will be no more corresponding memory leaks
in `paasta-api`.