APOLLO-565: Add a flag on updating haproxy on the map updates.
This commit adds a flag (--update-haproxy) which when provide updates
the haproxy with updated map information. The map file contains
the internal bridge IP address of a service and mesos task information
of the service.

This changes the present default behavior which updates haproxy on every
map update. This is not scalable (often fails with `Errno 11`)
and not needed since with synapse haproxy is restarted quite often.
The map_file is part of haproxy.cfg which means we don't need to
actively update haproxy, instead rely on these restarts for haproxy
to pick up updated info.

Note that we are only changing the default, hence, to revert to old behavior,
the flag (--update-haproxy) needs to be provided wherever it is invoked
(puppet).

Tested on a stage box.