use healthcheck_mode before mode
right now we default to looking at 'mode' in smartstack.yaml for
what mode should be used for healthchecks. there is nothing in place
that one must include mode for a service, and healthcheck mode 'should'
be enough. ensure that we look at 'healthcheck_mode' before 'mode' when
figuring out the healthcheck for a service.