Re-factor paasta cluster boost
I'd like to re-use the cluster boost logic to boost the service
autoscalers on a per-service level. Soon we will only need the
clusterman metrics bit for clusters anyway so I intend to separate that
block completely soon. This is just the refactor and I will add the
changes to the service autoscaler and boost cli in a second commit.

This:
* renames the module to load_boost.
* makes all the functions take a zk_boost_path which is {region}/{pool}
for cluster boost but can be like {cluster}/{service}/{instance} for the
service autoscaler
* adds some tests and tidy up