Increase agent statistics TTL
* increase the TTL to 30s for the statistics endpoint on the agent so
that the caller has a good chance of not having to re-retrieve that data

The motivation here is to speedup the service autoscaler which needs
this endpoint to get CPU stats. Experiments in stage show a 2x perf
boost with this commit.

Note: this only works because we already cache the `slave()` objects
when calling `stats()` on a `Task` object. If we didn't cache them then
each call to `slave()` would have its own cache.