Re-factor service autoscaler to take services list
This makes it possible to pass a list of services on the cli and will
make it possible for us to run several autoscalers to parallelise the
work. This moves the autoscaling lock to per service instance and uses
PaastaServiceConfigLoader in place of load_marathon_service_config.

By default this commit doesn't change anything. If you run without
arguments we will autsocale all services.