Add setup custom resource script
This is a script that allows us to specify a
file prefix like flinkcluster- or cassandra- etc. and map it to a
kubernetes custom object like FlinkCluster. For now it just adds the
neccessary metadata and submits the instances YAML config as the spec for
the k8s object. Like other paasta things we sha and check the sha to see
if we need to update.

Unit tests to follow if we think this is the way to go.