Creates an audit log across services
This commit creates an "audit log" of administrative actions taken on services
and clusters. The goal of this log is to provide a central view of actions
being taken to services across the platform. Currently this information is both
incompletely and inconsistently logged, and the information that is logged is
distributed across every services' individual logs, which is challenging and
slow to collate.

This change also renames the setting for the 'file' log writer that was
previously named 'delimeter' to 'delimiter', because spelling.