Bump pysensu-yelp, add more monitoring options
I'm planning on refactoring this soon to just pass through any supported
options to pysensu-yelp (except a few that can't be configured for
example). I'm hoping this should make thing simpler anyway since a new
function and dictionary key doesn't have to be added for every new
option in pysensu-yelp.