Enable registration of non-paasta k8s services
This makes it so we can register things from outside of paasta's k8s
namespaces. In addition you won't have to have a paasta instance config
file you just need to write the namespace you want to register to k8s
annotations. You still need a smartstack.yaml to define the settings for
the namespace you wish to register.

Motivation is so that operators outside of paasta can just write these
annotations if they want to register pods in smartstack.