Fix deployd bounce_length timer
We currently include bounces where we just run setup_marathon_job and
find that the state is already good. In our dashboards we mostly exclude
this by just excluding stuff less than 60s which is the default sleep
between reprocessing. That's not always the case though if the queues
are busy or we ever change that sleep. This should ditch the metric for
things that aren't processed more than once i.e. they aren't really
bounced.